requirements.txt
numpy==2.1.1
pyenchant==3.2.2
